+++ frontend/modules/module.py
@@ -8,6 +8,8 @@ Last Modified by: Hanyu Wang
 Last Modified time: 2024-06-25 23:55:15
 """
 
+from typing import List, Dict
+
 from .dfg import *
 from .ports import *
 from .definitions import *
@@ -17,26 +19,26 @@ class ModuleInst:
     def __init__(self, instName: str, moduleName: str) -> None:
         self.instName: str = instName
         self.moduleName: str = moduleName
-        self.assignment_list: list = []
-        self.parameter_list: dict = {}
+        self.assignment_list: List[Assignment] = []
+        self.parameter_list: Dict[str, str] = {}
 
-    def setAssignments(self, assignments: list):
+    def setAssignments(self, assignments: List[Assignment]) -> None:
         if not assignments:
             return
         self.assignment_list = assignments
 
-    def addParameters(self, parameters: dict):
+    def addParameters(self, parameters: Dict[str, str]) -> None:
         if not parameters:
             return
         self.parameter_list.update(parameters)
 
-    def addParameter(self, paramName: str, paramValue: str):
+    def addParameter(self, paramName: str, paramValue: str) -> None:
         self.parameter_list[paramName] = paramValue
 
-    def getParameters(self) -> dict:
+    def getParameters(self) -> Dict[str, str]:
         return self.parameter_list
 
-    def getAssignments(self) -> list:
+    def getAssignments(self) -> List[Assignment]:
         return self.assignment_list
 
     @property
@@ -65,10 +67,9 @@ class Module(Frame, Macros, ParameterHandler, BNGraph):
         Frame.__init__(self)
         ParameterHandler.__init__(self)
         BNGraph.__init__(self)
-        self.setName("top")
-        self.submodules = {}
-        self.macros = {}
+        self.submodules: Dict[str, ModuleInst] = {}
         self.node2assignment = {}
+        self.setName("top")
 
     def addInstance(self, inst: ModuleInst):
         self.submodules[inst.name] = inst
@@ -80,13 +81,17 @@ class Module(Frame, Macros, ParameterHandler, BNGraph):
         inst = self.submodules[instName]
         inst.addParameter(paramName, paramValue)
 
-    def setName(self, module_name: str):
+    def setName(self, module_name: str) -> None:
         self.module_name = module_name
 
-    def getName(self):
+    def getName(self) -> str:
         return self.module_name
 
-    def __repr__(self):
+    @property
+    def hasInstances(self) -> bool:
+        return len(self.submodules) > 0
+
+    def __repr__(self) -> str:
         return f"Module({self.module_name})"
 
     def __eq__(self, other: object) -> bool:
@@ -105,3 +110,20 @@ class Module(Frame, Macros, ParameterHandler, BNGraph):
 
     def __ne__(self, value: object) -> bool:
         return not self.__eq__(value)
+
+    def postProcess(self, params: dict = {}) -> None:
+        # turn submodules into assignments
+        if self.hasInstances:
+            # just skip it for now
+            return
+
+        # check the definitions
+        for port in self.getPortNames():
+            if self.isDefined(port) and self.getPort(port).isPi:
+                raise ValueError(
+                    f"Input port {port} is assigned in module {self.getName()}"
+                )
+            if not self.isDefined(port) and self.getPort(port).isPo:
+                raise ValueError(
+                    f"Output port {port} is not assigned in module {self.getName()}"
+                )