+++ tests/test_algorithms/test_tograph.py
@@ -4,8 +4,7 @@ from frontend import *
 def readAfterWrite(module: Module):
     graph = moduleToGraph(module)
     newModule = graphToModule(graph)
-    print(moduleToString(newModule))
-    assert module == newModule
+    return newModule
 
 
 def test_00_basic():
@@ -29,5 +28,29 @@ endmodule
     readAfterWrite(module)
 
 
+# Test 01
+# Mapping
+def test_01_mapping():
+    verilogStr = """
+module top (
+    input a,
+    output f
+);
+
+always @(*) begin
+    if (a) begin
+        f = 1;
+    end else begin
+        f = 0;
+    end
+end
+
+endmodule
+"""
+    moduleOld = transformVerilogToNetlist(verilogStr).getModuleAt(0)
+    moduleNew = readAfterWrite(moduleOld)
+    print(moduleToString(moduleNew))
+
+
 if __name__ == "__main__":
-    test_00_basic()
+    test_01_mapping()