+++ tests/test_writer.py
@@ -22,6 +22,8 @@ module top(
 input  a;
 input  b;
 output  f;
+
+assign f = a & b;
 endmodule
     """
     assert write_after_read(verilogString)
@@ -40,6 +42,9 @@ module top (
 );
     output c;
     input [1:0] d;
+    
+    assign f = d;
+    assign c = a & b;
 endmodule
 """
     assert write_after_read(verilogString)
@@ -521,9 +526,9 @@ module top (
 );
 
 add add_inst (
-    .a(a),
-    .b(b),
-    .f(f)
+    .a_in(a),
+    .b_in(b),
+    .f_out(f)
 );
 
 defparam add_inst.WIDTH = 4;
@@ -622,6 +627,28 @@ endmodule
     assert write_after_read(verilogString)
 
 
+# Test 31
+# Port Range
+def test_31_port_range():
+    verilogString = """
+module top (
+    input wire a,
+    input wire b,
+    output reg f
+);
+
+wire [3:0] a;
+wire [3:0] b;
+reg [3:0] f;
+always @(*) begin
+    f = a + b;
+end
+
+endmodule
+"""
+    assert write_after_read(verilogString)
+
+
 if __name__ == "__main__":
     # test_00_write_assignment()
     # test_01_write_assignment()
@@ -649,8 +676,9 @@ if __name__ == "__main__":
     # test_23_parameters()
     # test_24_parameters()
     # test_25_parameters()
-    # test_26_module_instantiation()
+    test_26_module_instantiation()
     # test_27_module_instantiation()
     # test_28_module_instantiation()
     # test_29_module_instantiation()
-    test_30_port_type()
+    # test_30_port_type()
+    # test_31_port_range()