+++ docs/source/changelog.rst
@@ -1,6 +1,12 @@
 Change log
 ==========
 
+v0.1.6 (Not Yet Released)
+-------------------------
+
+* Graph Representation
+    - Update class name `DFGNode` to `BNode`, to separate the class from the CDFG
+
 v0.1.4 (September 27, 2024)
 ---------------------------
 
@@ -11,7 +17,7 @@ v0.1.4 (September 27, 2024)
 * Verilog I/O:
     - Add support for module instantiation in SystemVerilog
     - Add syntax checking for port type, directions, and ranges (as well as the test cases, see: `test/test_ports.py` and `test/test_ports_false.py`). 
-    - Update many classes to `dataclass` (e.g. `Port`, `Module`, `Netlist`, `Assignment`, `DFGNode`, `OPNode`, `OPType`, `BNGraph`) to make them immutable
+    - Update many classes to `dataclass` (e.g. `Port`, `Module`, `Netlist`, `Assignment`, `BNode`, `OPNode`, `OPType`, `BNGraph`) to make them immutable
 
 v0.1.2 (September 22, 2024)
 ---------------------------
@@ -30,8 +36,8 @@ v0.1.0 (September 22, 2024)
 ---------------------------
 
 * Graph Representation
-    - Add `DFGNode` class to represent nodes in a dataflow graph
-    - Add `OPNode` class as a child class of `DFGNode` to represent operation nodes
+    - Add `BNode` class to represent nodes in a dataflow graph
+    - Add `OPNode` class as a child class of `BNode` to represent operation nodes
     - Add `OPType` class to represent the type of operation nodes
     - Add `BNGraph` class